requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
albumentations>=1.0.0
numpy>=1.21.0
tqdm>=4.62.0
PyYAML>=5.4.0
wandb>=0.12.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
